requirements.txt
openai>=1.3.0
datasets>=2.14.0
pandas>=2.0.0
pillow>=10.0.0
requests>=2.31.0 